requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1